requirements_local.txt
moto
boto3
flake8
textract
PyMuPDF
requests
pytest-mock